requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4
numpy==1.26.2

# CLI
click==8.1.7

# NLP (Week 2)
spacy==3.7.2
textblob==0.17.1
openai==1.3.7

# Reporting (Week 5)
reportlab==4.0.7
matplotlib==3.8.2
plotly==5.18.0

# Dashboard (Week 6)
streamlit==1.29.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-benchmark==4.0.0
pytest-xdist==3.5.0  # Parallel test execution
pytest-timeout==2.2.0  # Test timeouts

# Security
safety==2.3.5
bandit==1.7.5

# Monitoring
psutil==5.9.6  # System metrics

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Apify Integration
apify-client==1.7.0

